@@ -2,15 +2,31 @@
 
 namespace Ruvents\TwigExtensions;
 
-class SortExtension extends \Twig_Extension
+use Symfony\Component\PropertyAccess\PropertyAccess;
+use Symfony\Component\PropertyAccess\PropertyAccessorInterface;
+use Twig\Error\RuntimeError;
+use Twig\Extension\AbstractExtension;
+use Twig\TwigFilter;
+
+class SortExtension extends AbstractExtension
 {
+    /**
+     * @var PropertyAccessorInterface
+     */
+    private $propertyAccessor;
+
+    public function __construct(PropertyAccessorInterface $propertyAccessor = null)
+    {
+        $this->propertyAccessor = $propertyAccessor ?: PropertyAccess::createPropertyAccessor();
+    }
+
     /**
      * {@inheritdoc}
      */
     public function getFilters()
     {
         return [
-            new \Twig_SimpleFilter('sort', function ($traversable, $flags = null, $preserveKeys = true) {
+            new TwigFilter('sort', function ($traversable, $flags = null, $preserveKeys = true) {
                 $array = $this->toArray($traversable);
 
                 if ($preserveKeys) {
@@ -21,17 +37,33 @@ class SortExtension extends \Twig_Extension
 
                 return $array;
             }),
-            new \Twig_SimpleFilter('ksort', function ($traversable, $flags = null) {
+            new TwigFilter('ksort', function ($traversable, $flags = null) {
                 $array = $this->toArray($traversable);
 
                 ksort($array, $flags);
 
                 return $array;
             }),
-            new \Twig_SimpleFilter('natsort', function ($traversable) {
+            new TwigFilter('natsort', function ($traversable) {
                 $array = $this->toArray($traversable);
 
                 natsort($array);
+
+                return $array;
+            }),
+            new TwigFilter('sort_by', $sortBy = function ($array, $path, $preserveKeys = true) {
+                $function = $preserveKeys ? 'uasort' : 'usort';
+
+                $function($array, function ($a, $b) use ($path) {
+                    $aValue = $this->propertyAccessor->getValue($a, $path);
+                    $bValue = $this->propertyAccessor->getValue($b, $path);
+
+                    if ($aValue == $bValue) {
+                        return 0;
+                    }
+
+                    return $aValue > $bValue ? 1 : -1;
+                });
 
                 return $array;
             }),
@@ -39,18 +71,10 @@ class SortExtension extends \Twig_Extension
     }
 
     /**
-     * {@inheritdoc}
-     */
-    public function getName()
-    {
-        return 'ruvents_twig_extensions.sort';
-    }
-
-    /**
      * @param array|\Traversable $traversable
      *
      * @return array
-     * @throws \Twig_Error_Runtime
+     * @throws RuntimeError
      */
     private function toArray($traversable)
     {
@@ -60,7 +84,7 @@ class SortExtension extends \Twig_Extension
             return iterator_to_array($traversable);
         }
 
-        throw new \Twig_Error_Runtime(sprintf(
+        throw new RuntimeError(sprintf(
             'Sorted variable must be an array or an instance of Traversable, got "%s".',
             gettype($traversable)
         ));
