@@ -2,7 +2,10 @@
 
 namespace Ruvents\TwigExtensions;
 
-class StringExtension extends \Twig_Extension
+use Twig\Extension\AbstractExtension;
+use Twig\TwigFilter;
+
+class StringExtension extends AbstractExtension
 {
     /**
      * {@inheritdoc}
@@ -10,22 +13,12 @@ class StringExtension extends \Twig_Extension
     public function getFilters()
     {
         return [
-            new \Twig_SimpleFilter('clear_tags', $clearTags = function ($string) {
+            new TwigFilter('clear_tags', $clearTags = function ($string) {
                 return preg_replace('/<([a-z][a-z0-9]*)[^>]*?(\/?)>/i', '<$1$2>', $string);
             }),
-            new \Twig_SimpleFilter('cleartags', $clearTags,
-                ['deprecated' => true, 'alternative' => 'clear_tags']),
-            new \Twig_SimpleFilter('respace', function ($string) {
+            new TwigFilter('respace', function ($string) {
                 return preg_replace("/[\s\p{Z}]+/ui", ' ', $string);
             }),
         ];
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function getName()
-    {
-        return 'ruvents_twig_extensions.string';
     }
 }
