@@ -3,8 +3,10 @@
 namespace Ruvents\TwigExtensions;
 
 use Doctrine\Common\Inflector\Inflector;
+use Twig\Extension\AbstractExtension;
+use Twig\TwigFilter;
 
-class InflectorExtension extends \Twig_Extension
+class InflectorExtension extends AbstractExtension
 {
     /**
      * {@inheritdoc}
@@ -12,26 +14,18 @@ class InflectorExtension extends \Twig_Extension
     public function getFilters()
     {
         return [
-            new \Twig_SimpleFilter('underscorize', function ($string) {
+            new TwigFilter('underscorize', function ($string) {
                 return Inflector::tableize($string);
             }),
-            new \Twig_SimpleFilter('classify', function ($string) {
+            new TwigFilter('classify', function ($string) {
                 return Inflector::classify($string);
             }),
-            new \Twig_SimpleFilter('camelize', function ($string) {
+            new TwigFilter('camelize', function ($string) {
                 return Inflector::camelize($string);
             }),
-            new \Twig_SimpleFilter('ucwords', function ($string, $delimiters = " \n\t\r\0\x0B-") {
+            new TwigFilter('ucwords', function ($string, $delimiters = " \n\t\r\0\x0B-") {
                 return Inflector::ucwords($string, $delimiters);
             }),
         ];
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function getName()
-    {
-        return 'ruvents_twig_extensions.inflector';
     }
 }
